Update and rename auto_pages_v0.3.JSX to auto_pages_v0.5.JSX
Updated the script to version 0.5 so that it now uses customizable dialog windows (provided by ScriptUI) instead of the built-in ones in InDesign, that allows me to add custom buttons. To do for the future: add functionality to the "Revert" button. There are currently problems with the for-loop in applyMasterPages (it breaks in the middle of the first iteration). Need to figure out why and fix it.